pyproject.toml
[tool.poetry]
name = "deltaupgrade"
version = "0.1.0"
description = ""
authors = ["ttran <ms.tam.h.tran@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.22.0"
python-dotenv = "^1.0.1"
webdriver-manager = "^4.0.1"
undetected-chromedriver = "^3.5.5"
setuptools = "^70.3.0"
wheel = "^0.43.0"
certifi = "^2024.7.4"
twilio = "^9.2.3"
google-api-python-client = "^2.137.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
